requirements.txt
absl-py==1.4.0
asttokens==2.2.1
pyxlsb==1.0.10
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
blis==0.7.9
breadability==0.1.20
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
PyGithub
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
funcy==2.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.6
lxml==4.9.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pyxlsb==1.0.10
PyYAML==6.0.1
pyzmq==25.1.0
rake-nltk==1.0.6
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.3.3
segtok==1.5.11
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
streamlit==1.27.2
sumy==0.11.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.32.1
typer==0.7.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.16
validators==0.22.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
wordcloud==1.9.2
wrapt==1.15.0
XlsxWriter==3.1.6
yake==0.4.8
zipp==3.17.0
